+++ NEWS.md
@@ -13,6 +13,9 @@ ethos 0.2.3 prerelease
   treated as a valid code point.
 - Fixes issue in the parser which did not guard for overflow of 32 bit unsigned
   values.
+- Fixes an issue where non-ground nil terminators would not be properly computed for list construction operators with types where the element type is different from the return type of the operator.
+- Fixes for evaluation of `:left-assoc-nil` and `:left-assoc-non-singleton-nil` operators.
+- Variables `(eo::var s T)` are now considered ordinary terms, which are hence allowed in patterns.
 
 
 ethos 0.2.2